requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.4.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.4=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py39hdda0065_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=2.0.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.2.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.4=py39h06a4308_0
importlib-resources=6.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.2=pypi_0
ipython=8.16.1=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=py39h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jupyter-client=8.3.1=pypi_0
jupyter-core=5.3.2=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.8.0=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numpy=1.26.0=py39h5f9d8c6_0
numpy-base=1.26.0=py39hb5e798b_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.11=h7f8727e_2
packaging=23.2=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.1=py39ha6cbd5a_0
pip=23.2.1=py39h06a4308_0
platformdirs=3.11.0=pypi_0
prompt-toolkit=3.0.39=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-graphviz=0.20.1=pypi_0
pytorch=2.1.0=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-metric-learning=2.4.1=pypi_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=py39h5eee18b_1
pyzmq=25.1.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard-logger=0.1.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.1.0=py39_cu118
torchtriton=2.1.0=py39
torchvision=0.16.0=py39_cu118
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.11.2=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py39h06a4308_0
wcwidth=0.2.8=pypi_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
